modules/module-08-streaming-basics/requirements.txt
# Module 08: Streaming Basics - Requirements

# Core Dependencies
kafka-python==2.0.2
confluent-kafka==2.3.0
apache-flink==1.18.0
pyflink==1.18.0

# Data Processing
pandas==2.0.3
pyarrow==14.0.1
avro-python3==1.10.2
fastavro==1.9.0

# AWS Services
boto3==1.28.85
aioboto3==12.0.0
aws-kinesis-agg==1.2.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-timeout==2.2.0
moto==4.2.9

# Validation & Quality
great-expectations==0.18.3
pydantic==2.5.0

# Monitoring & Logging
structlog==23.2.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
click==8.1.7

# Development
ipython==8.18.1
jupyter==1.0.0
black==23.11.0
flake8==6.1.0
